# Performance optimization for MultiThreadedStrategy