--- a/core/include/gnuradio-4.0/Buffer.hpp
+++ b/core/include/gnuradio-4.0/Buffer.hpp
@@ -18,12 +18,11 @@ struct first_template_arg_helper<TemplateType<ValueType, OtherTypes...>> {
 };
 
 template<typename T>
-constexpr auto *
-value_type_helper() {
+constexpr auto* value_type_helper() {
     if constexpr (requires { typename T::value_type; }) {
-        return static_cast<typename T::value_type *>(nullptr);
+        return static_cast<typename T::value_type*>(nullptr);
     } else {
-        return static_cast<typename first_template_arg_helper<T>::value_type *>(nullptr);
+        return static_cast<typename first_template_arg_helper<T>::value_type*>(nullptr);
     }
 }
 
@@ -47,10 +46,10 @@ static_assert(std::is_same_v<double, value_type_t<std::array<double, 42>>>);
 
 // TODO: find a better place for these 3 concepts
 template<typename From, typename To>
-concept convertible_from_lvalue_to = std::convertible_to<const std::remove_cvref_t<From> &, To>;
+concept convertible_from_lvalue_to = std::convertible_to<const std::remove_cvref_t<From>&, To>;
 
 template<typename From, typename To>
-concept convertible_from_rvalue_to = std::convertible_to<std::remove_cvref_t<From> &&, To>;
+concept convertible_from_rvalue_to = std::convertible_to<std::remove_cvref_t<From>&&, To>;
 
 template<typename From, typename To>
 concept convertible_from_lvalue_only = convertible_from_lvalue_to<From, To> && !convertible_from_rvalue_to<From, To>;
@@ -65,16 +64,16 @@ template<typename T>
 concept ConstSpanLvalueLike = convertible_from_lvalue_only<T, std::span<const std::remove_cvref_t<typename T::value_type>>>;
 
 template<typename T>
-concept ConsumableSpan = std::ranges::contiguous_range<T> && ConstSpanLvalueLike<T> && requires(T &s) { s.consume(0); };
+concept ConsumableSpan = std::ranges::contiguous_range<T> && ConstSpanLvalueLike<T> && requires(T& s) { s.consume(0); };
 
 template<typename T>
-concept PublishableSpan = std::ranges::contiguous_range<T> && std::ranges::output_range<T, std::remove_cvref_t<typename T::value_type>> && SpanLike<T> && requires(T &s) { s.publish(0UZ); };
+concept PublishableSpan = std::ranges::contiguous_range<T> && std::ranges::output_range<T, std::remove_cvref_t<typename T::value_type>> && SpanLike<T> && requires(T& s) { s.publish(0UZ); };
 
 // clang-format off
 // disable formatting until clang-format (v16) supporting concepts
 template<class T>
-concept BufferReader = requires(T /*const*/ t, const std::size_t n_items) {
-    { t.get(n_items) } ; // TODO: get() returns CircularBuffer::buffer_reader::ConsumableInputRange
+concept BufferReaderLike = requires(T /*const*/ t, const std::size_t nItems) {
+    { t.get(nItems) } ; // TODO: get() returns CircularBuffer::Reader::ConsumableInputRange
     { t.position() }       -> std::same_as<std::make_signed_t<std::size_t>>;
     { t.available() }      -> std::same_as<std::size_t>;
     { t.buffer() };
@@ -82,27 +81,21 @@ concept BufferReader = requires(T /*const*/ t, const std::size_t n_items) {
     { t.isConsumeRequested()} -> std::same_as<bool>;
 };
 
-template<class Fn, typename T, typename ...Args>
-concept WriterCallback = std::is_invocable_v<Fn, std::span<T>&, std::make_signed_t<std::size_t>, Args...> || std::is_invocable_v<Fn, std::span<T>&, Args...>;
-
 template<class T, typename ...Args>
-concept BufferWriter = requires(T t, const std::size_t n_items, std::pair<std::size_t, std::make_signed_t<std::size_t>> token, Args ...args) {
-    { t.publish([](std::span<util::value_type_t<T>> &/*writable_data*/, Args ...) { /* */ }, n_items, args...) }                                 -> std::same_as<void>;
-    { t.publish([](std::span<util::value_type_t<T>> &/*writable_data*/, std::make_signed_t<std::size_t> /* writePos */, Args ...) { /* */  }, n_items, args...) }   -> std::same_as<void>;
-    { t.try_publish([](std::span<util::value_type_t<T>> &/*writable_data*/, Args ...) { /* */ }, n_items, args...) }                             -> std::same_as<bool>;
-    { t.try_publish([](std::span<util::value_type_t<T>> &/*writable_data*/, std::make_signed_t<std::size_t> /* writePos */, Args ...) { /* */  }, n_items, args...) }-> std::same_as<bool>;
-    { t.reserve(n_items) };// TODO: reserve() returns CircularBuffer::buffer_writer::PublishableOutputRange
+concept BufferWriterLike = requires(T t, const std::size_t nItems) {
+    { t.tryReserve(nItems) };// TODO: tryReserve() returns CircularBuffer::Writer::PublishableOutputRange
+    { t.reserve(nItems) };// TODO: reserve() returns CircularBuffer::Writer::PublishableOutputRange
     { t.available() }         -> std::same_as<std::size_t>;
     { t.buffer() };
     { t.nSamplesPublished()} -> std::same_as<std::size_t>;
 };
 
 template<class T, typename ...Args>
-concept Buffer = requires(T t, const std::size_t min_size, Args ...args) {
+concept BufferLike = requires(T t, const std::size_t min_size, Args ...args) {
     { T(min_size, args...) };
     { t.size() }       -> std::same_as<std::size_t>;
-    { t.new_reader() } -> BufferReader;
-    { t.new_writer() } -> BufferWriter;
+    { t.new_reader() } -> BufferReaderLike;
+    { t.new_writer() } -> BufferWriterLike;
 };
 
 // compile-time unit-tests
@@ -116,16 +109,9 @@ template <typename T, typename... Args>
 using NoSequenceParameter = decltype([](std::span<T>&, Args...) { /* */ });
 } // namespace test
 
-static_assert(!Buffer<test::non_compliant_class<int>>);
-static_assert(!BufferReader<test::non_compliant_class<int>>);
-static_assert(!BufferWriter<test::non_compliant_class<int>>);
-
-static_assert(WriterCallback<test::WithSequenceParameter<int>, int>);
-static_assert(!WriterCallback<test::WithSequenceParameter<int>, int, std::span<bool>>);
-static_assert(WriterCallback<test::WithSequenceParameter<int, std::span<bool>>, int, std::span<bool>>);
-static_assert(WriterCallback<test::NoSequenceParameter<int>, int>);
-static_assert(!WriterCallback<test::NoSequenceParameter<int>, int, std::span<bool>>);
-static_assert(WriterCallback<test::NoSequenceParameter<int, std::span<bool>>, int, std::span<bool>>);
+static_assert(!BufferLike<test::non_compliant_class<int>>);
+static_assert(!BufferReaderLike<test::non_compliant_class<int>>);
+static_assert(!BufferWriterLike<test::non_compliant_class<int>>);
 // clang-format on
 } // namespace gr